buildコマンドを明示化

+++ scripts/uploader.js
@@ -7,8 +7,9 @@ const args  = process.argv.slice(2)
 const watch = args.includes('--watch') || args.includes('-W')
 
 const deploy = () => {
-  return new Promise((resolve, reject) => {
-    let command = `yarn kintone-customize-uploader ${outputManifest(env, watch)}`
+  return new Promise(async (resolve, reject) => {
+    const file = await outputManifest(env, watch)
+    let command = `yarn kintone-customize-uploader ${file}`
     const options = [
       ['--base-url', env.KINTONE_BASE_URL],
       ['--username', env.KINTONE_USERNAME],
@@ -31,33 +32,45 @@ module.exports =  deploy
 
 
 function outputManifest(env, watch) {
-  const json = {
-    app: env.APP_ID,
-    scope: env.APP_SCOPE,
-  }
-  if (env.USE_DESKTOP) {
-    json.desktop = watch ? {
-      js: isExists('app.min.js') ? [ `https://localhost:3000/dist/app.min.js` ] : [],
-      css: isExists('app.min.css') ? [ `https://localhost:3000/dist/app.min.css` ] : [],
-    } : {
-      js: isExists(`${env.OUT_FILENAME || 'app'}.min.js`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.js` ] : [],
-      css: isExists(`${env.OUT_FILENAME || 'app'}.min.css`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.css` ] : [],
-    }
-  }
-  if (env.USE_MOBILE) {
-    json.mobile = watch ? {
-      js: isExists('app.min.js') ? [ `https://localhost:3000/dist/app.min.js` ] : [],
-      css: isExists('app.min.css') ? [ `https://localhost:3000/dist/app.min.css` ] : [],
-    } : {
-      js: isExists(`${env.OUT_FILENAME || 'app'}.min.js`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.js` ] : [],
-      css: isExists(`${env.OUT_FILENAME || 'app'}.min.css`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.css` ] : [],
-    }
-  }
+  return new Promise((resolve, reject) => {
+    const fileExist = setInterval(() => {
+      if (
+        isExists('app.min.js') || isExists('app.min.css') ||
+        isExists(`${env.OUT_FILENAME || 'app'}.min.js`) ||
+        isExists(`${env.OUT_FILENAME || 'app'}.min.css`)
+      ) {
+        const json = {
+          app: env.APP_ID,
+          scope: env.APP_SCOPE,
+        }
+        if (env.USE_DESKTOP) {
+          json.desktop = watch ? {
+            js: isExists('app.min.js') ? [ `https://localhost:3000/dist/app.min.js` ] : [],
+            css: isExists('app.min.css') ? [ `https://localhost:3000/dist/app.min.css` ] : [],
+          } : {
+            js: isExists(`${env.OUT_FILENAME || 'app'}.min.js`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.js` ] : [],
+            css: isExists(`${env.OUT_FILENAME || 'app'}.min.css`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.css` ] : [],
+          }
+        }
+        if (env.USE_MOBILE) {
+          json.mobile = watch ? {
+            js: isExists('app.min.js') ? [ `https://localhost:3000/dist/app.min.js` ] : [],
+            css: isExists('app.min.css') ? [ `https://localhost:3000/dist/app.min.css` ] : [],
+          } : {
+            js: isExists(`${env.OUT_FILENAME || 'app'}.min.js`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.js` ] : [],
+            css: isExists(`${env.OUT_FILENAME || 'app'}.min.css`) ? [ `dist/${env.OUT_FILENAME || 'app'}.min.css` ] : [],
+          }
+        }
 
-  const manifest = './env/manifest.json'
-  console.log(json)
-  fs.writeFileSync(manifest, JSON.stringify(json, null, "\t"))
-  return manifest
+        const manifest = './env/manifest.json'
+        fs.writeFileSync(manifest, JSON.stringify(json, null, "\t"))
+        resolve(manifest)
+
+        clearInterval(fileExist)
+      }
+    }, 500)
+
+  })
 }
